Add new state for when Jenkins is unavailable
If Jenkins was down the scripts would set the lights for
"SUCCESS", which is sort of misleading. This commit
introduces a new state, JENKINS_DOWN, that shuts a light
off if the script is unable to connect to Jenkins due
to Jenkins being down. All other states have been updated
to ensure the light is turned back on when a successful
Jenkins response happens.